# slices: optimize Compact and CompactFunc